requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
pillow==10.4.0
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
twilio==9.3.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.13.1
